scifio-ify project
Changed package prefixes, artifact groupid, etc.
