+++ src/test/java/io/scif/labeling/LabelingIOTest.java
@@ -31,13 +31,13 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imglib2.labeling;
+package io.scif.labeling;
 
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
+import io.scif.labeling.data.Container;
 import net.imglib2.img.Img;
 import net.imglib2.img.array.ArrayImgs;
-import net.imglib2.labeling.data.Container;
 import net.imglib2.labeling.data.LabelingData;
 import net.imglib2.roi.labeling.ImgLabeling;
 import net.imglib2.type.numeric.integer.IntType;
