+++ src/test/java/io/scif/labeling/tutorials/E01_LoadLabeling.java
@@ -31,12 +31,10 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imglib2.labeling.tutorials;
+package io.scif.labeling.tutorials;
 
-import net.imglib2.labeling.LabelingIOService;
-import net.imglib2.labeling.data.Container;
-import net.imglib2.labeling.data.LabelingData;
-import net.imglib2.labeling.data.TypeTokenWrapper;
+import io.scif.labeling.LabelingIOService;
+import io.scif.labeling.data.Container;
 import net.imglib2.roi.labeling.ImgLabeling;
 import net.imglib2.type.numeric.integer.IntType;
 import org.junit.Assert;
@@ -45,7 +43,6 @@ import org.junit.Test;
 import org.scijava.Context;
 
 import java.io.IOException;
-import java.lang.reflect.Type;
 
 public class E01_LoadLabeling {
